The source and target levels are javac options and the java compiler does not need to be exposed.
Summary:
We previously modeled these as being part of the compiler, which is confusing.

In addition the JavaCompilerEnvironment was being passed around since it was
required to build the JavacOptions. Since the JavacOptions are now being
passed in from above, there is no longer any need to pass around the
JavaCompilerEnvironment directly. Because of this, we now no longer expose
this information from the JavaBuckConfig.

Test Plan: buck test --all