New Hamcrest matcher hasConsecutiveItems()
Summary:
Hamcrest includes `Matcher` implementations which:

1) Check if a collection includes a single item anywhere (`hasItem`)
2) Check if a collection includes multiple items anywhere (`hasItems`)
3) Check if a collection exactly matches another collection in order (`contains`)
4) Check if a collection exactly matches another collection in any order (`containsInAnyOrder`)

Bizarrely, Hamcrest even recently added a `Matcher` which searches for
the items of a collection in relative order anywhere inside the target
collection (`containsInRelativeOrder`). (Note that this searches anywhere
inside the container, unlike the other `contains` `Matcher`s!)

https://github.com/hamcrest/JavaHamcrest/pull/74

We need a `Matcher` which finds a sub-collection inside a larger
collection in order (i.e., finding `[2, 3]` inside `[1, 2, 3]` but not
inside `[1, 2, 42, 3]`).

This diff adds that matcher, `hasConsecutiveItems()`, along with tests.

Test Plan: Unit tests includes.