Easy: Fix DaemonIntegrationTest
Summary:
This is the same fix as before. Silly me for thinking
we wrote tests which didn't depend on the host environment! Now we copy
the host environment into the test.

Test Plan:
`ant java-test -Dtest.class=DaemonIntegrationTest`
before and after change. Before, it failed with `Unexpected exit code. expected:<0> but was:<1>`.
Now, it passes.