Pass architecture, platform, and SDK version to AppleCxxPlatform
Summary:
To compile for Apple platforms, we need clients to specify
the architecture, platform (iPhone / OS X), and SDK version.

This adds extra parameters to `AppleCxxPlatform` to specify these
values. I'm not sure if I should use `AppleSdk` here or
not: that object currently includes *all* supported architectures and
the *maximum* supported SDK version, but when compiling, we want a single
architecture and the *minimum* SDK version.

Test Plan: Unit test updated.