Flip chained method order.
Summary:
`Path#getParent` on a path that only has a file name returns null instead
of the parent path. Flip the order of the method calls around so we first
convert the path to absolute, then get the parent.
This will still fail if the BUCK file is in the filesystem root.

Test Plan: buck audit owner .buckconfig in a project that has a BUCK file in the root directory