Use a filtered environment to run integration tests
Summary:
@bgertzfield fixed buckd so that it determined whether to format output
for an interactive terminal based on the clients context.  However,
several of our integration tests rely the output of their runs coming
from a non-interactive run.  Since our integration tests inherit the
full environment of the client, the special `NAILGUN_TTY_*` environment
variables propagate through when running tests using buckd and from
a TTY, which causes the integration tests to emit colored output, and
breaks several of the above tests.

This diff filters most of the environment we run the integration tests
with, preventing this issue.  Eventually it'd be nice to run them
under a blank environment, but we currently rely on a few vars.

Test Plan: `buck test --all` form a TTY.