C/C++: use propagated preprocessor flags locally
Summary:
This applies propagated preprocessor flags to local compilations
as well, which mimics the behavior we use internally.  Generally,
the propagated preprocessor flag option is used when a preprocessor
flag needs to be attached to the headers a library exports, which
mean we should use it for local compilation and for compilation of
dependents.

(We never seen a case where we *only* want a preprocessor flag to
be used by dependents, but not by the current rule itself.)

Test Plan: added integration test