Start to break the assumption that a TargetGraph can only transform to an ActionGraph.
Summary:
The forthcoming `buck fetch` command needs to transform the TargetGraph
in a subtly different way to the current way we do to construct what is
now the ActionGraph. We've discussed the hypothetical need to construct
a different graph from the ActionGraph before. We have now reached that
point.

This change introduces the concept of the TargetGraphTransformer, which
is responsible for this conversion. It is kept deliberately minimal for
ease of review.

Test Plan: buck test --all