Correctly set headermaps if tested source is a bundle.
Summary:
If a bundle specifies the `tests` parameter, we want the specified tests to have
access to the header files of the library included in the bundle. Before this
diff this would only work if the `tests` parameter was specified on the library
itself. This diff makes sure to check both the value specified for the library
and for the bundle.

Test Plan:
  * A new unit test.
  * Create a project that uses headermaps and has a library inside a bundle
    where the bundle specifies the tests. Check that the test has access to
    the private headers of the library.