allow collecting dependencies to take multiple args
Summary:
In order to support generating combined tests, which will require grabbing
dependencies from multiple library nodes, the collection functions are
refactored to take in a list of nodes as input instead of a single one.

Some functions only return values for ancestor instead of ancestor-or-self, for
these I added a flag for returning values for the passed in nodes as well.

Performance impact is minimal.

Test Plan:
existing tests pass
output should be identical