Don't use action nodes to create shell script phases.
Summary:
Shell script phases in the Xcode project generation are created using action
nodes instead of target nodes. This is more complicated, but is needed to
support the more advanced features such as dependencies, caching and expanding
target references. Unfortunately all these feature are partly unimplemented
and partly impossible to replicate inside of Xcode, so there is no point in
using the action nodes here. This changes the creation of shell script phases
to use the target nodes instead.

Test Plan: CI, generate a project with a shell script phase