Loosen default platform requirement in AppleSdkDiscovery
Summary:
A few users have a partial install of Xcode which is
missing platforms. This caused the `Preconditions` check I added
to `AppleSdkDiscovery` to fire.

This loosens the requirement to simply return no SDKs in this case,
and adds a unit test to confirm the behavior.

Test Plan: Unit test included.