remove BuildTarget from NewNativeTargetProjectMutator
Summary:
Removes the need to pass in `BuildTarget` for `NewNativeTargetProjectMutator`.

This required refactoring converting framework strings into file references
earlier, in the ProjectGenerator.

In the process I fixed one probable bug, the strings which are not
SourceTreePaths were resolved against the incorrect project.

This is needed in order to avoid synthesizing build targets when generating new
test targets from scratch.

Test Plan: CI