Fat JARs: package native libraries into JARs
Summary:
This diff adds support for Java rules to depend on C/C++ rules.
The result is that top-level `java_binary` rules generate a new
type of JAR: a "fat" JAR which packages up all the C/C++ shared
library dependencies, and uses a preamble to unpack these shared
libs to a temp location which is exported to the system library
environment variable so that the system loader can find them when
they get loaded by the JAR.  The preamble then executes the real
main class in a subprocess, so that it uses the updated system
library paths.

Test Plan: added integration tests