Make sure python_test runs outside of the Buck project.
The `python_test` rules depended on a system property ("buck.home") that
was never set. As such, any time a `python_test` ran outside of the main
buck tree, it would fall back to `user.dir`. Obviously, this never would
not exist in a majority of projects, causing `python_test` to fail.

Tested manually on the gerrit repo after a version bump.