Always place combined project into the root group of its workspace
Summary:
Combined project is the only project in its workspace so it can always be placed
directly in its root. This add an option to override the path in the workspace
where a project gets placed and uses it to place combined projects to their
workspaces' roots.

Test Plan: Generate and compile an app as a combined project.