[buck-ios] Always compile tests as static libraries if groupable
Summary:
Due to the fact that projects are shared amongst all the generated workspaces,
tests which do not combine with anything in one workspace may still combine
with something else in another workspace. This necessitates that any test that
can be combined should be built as static libraries such that they will be
combined, instead of the previous behavior of only building them as static
libraries if they can be combined with a test of a given workspace.

Test Plan: modified unit tests to check this case.