Add DroidCon NYC 2014 presentation to the list of talks in the public docs.
Summary:
Because it has been over three months and Touch Lab has failed to
post videos for any of the talks for DroidCon NYC, it seems the best
way to get the information out there is to publish our slide deck
with detailed speaker notes.

There was also a live demo, which I redid today and recorded via QuickTime.
Because I am not smart enough to figure out how to serve a video in a way
that is reliably cross-browser and cross-platform, I decided to just upload it to
YouTube instead: https://www.youtube.com/watch?v=rNiZOYYmFcI#t=16.
This video is embedded on slide 3 of the presentation.

I updated the speaker notes for all of my slides. I cleaned up the few
notes that David had for his slides, but it would be best if he followed
up and added all of his speaker notes himself.

Test Plan:
http://localhost:9811/presentations/

Everything works except for the serving of the .svg file.
I believe this is only an issue locally because of a bug in soyweb:
https://github.com/bolinfest/plovr/blob/master/src/org/plovr/soy/server/RequestHandlerSelector.java

Once this is hosted on GitHub pages, I believe it should be fine.