Hashing for parsed build target rules
Summary:
To allow continuous integration systems to determine if a
build target has changed, we'd like a target graph-only alternative to
the rule key (since that depends on the action graph, and we don't want
to expose that).

This adds a `LoadingCache` to `Parser` which maps `BuildTarget`s to
`HashKey`s. Internally, it simply hashes the raw JSON objects parsed
out of `BUCK` files.

This does mean this implementation is dependent on the order of the keys
in the `BUCK` file.

Test Plan: Unit test included.