Easy: Relax FakeProjectFilesystem.walkRelativeFileTree()
Summary:
Previously, `FakeProjectFilesystem.walkRelativeFileTree` would
error out if passed a `Path` that refered to a file and not a directory.

This isn't how the real `ProjectFilesystem` works (you can pass it a `Path`
that refers to a file and it "walks" over just that file), so this makes
`FakeProjectFilesystem` behave the same as `ProjectFilesystem` and adds
a test.

Test Plan:
Unit test included. Ran against future diff that depended on
this behavior.