Test environments also need to include PATHEXT on Windows
Summary:
PATHEXT is normally empty on unixy platforms and the code defaults to
the empty string when the environment variable isn't found. On Windows though,
we need to have this or else we can't find the Python interpreter.

Test Plan:
'ant test' and testOrigFileIsIgnoredByAapt no longer fails due to
not being able to find a python binary when Python is on the path.