ProGuardObfuscateStep: always generate mapping.txt file
Summary:
Previously, if a user provided config had `-dontobfuscate` in it,
proguard wouldn't actually generate a mapping.txt file.  This led
to a gross exception hitting the user, since we record it via
`buildableContext.recordArtifact`.

This diff makes sure to always touch that file, so that we at least
always generate an empty one.

Test Plan: added test