Update code coverage check to ignore immutables.
Summary:
As we convert more of our codebase to use immutables the test coverage drops as
the generated classes have a lot of methods that are never called. This adds a
hack to the coverage verification script to ignore classes whose name starts
with "Immutable", which probably means they have been generated by immutables.
This also renames ImmutableDirectedAcyclicGraph so that doesn't get ignored.

Test Plan:
    buck test --all --code-coverage --code-coverage-format xml --no-results-cache
    python scripts/assert_code_coverage.py

Before and after:

    TOTAL                                     71.40%  75.20%  61.17%  74.31%  87.59%    6541
    TOTAL                                     73.70%  77.33%  63.65%  77.46%  88.10%    5697