HttpArtifactCache: use chunked streaming mode
Summary:
By default, the `URLConnection` class buffers the entire contents of
the HTTP request in memory so that it can set the `Content-Length`
header appropriately.  This diff enables "chunked" mode so that we
actually stream the artifact.

Test Plan:
I lowered the heap size buck uses to 100M, set my local checkout to
write to the HTTP cache, then created a genrule which just copied a 1G
chunk of random data and ran it, which repros the issue easily and
also verifies that this fix works.

It'd be nice to write an actual unittest to vet this but I'm not really
sure there's an easy way to do this.  Open to suggestions here.