Fix non-deterministic target hashing
Summary:
The JSON parser we use parses JSON objects into `HashMap`s, whose traversal
order depends on the hashes of the stored keys, which are not guaranteed to
be stable across launches or Java versions. That means that with the current
implementation, target hashes might not be stable. To fix that, instead of
traversing the provided map in its default traversal order, we will copy the
key-value pairs into an `ImmutableSortedMap`, which will ensure that we will
traverse the elements in lexicographical order of their `String` keys.

Test Plan:
  * `buck test --all` + a new unit test.
  * Running the related tests under Java 8 now succeeds.