Index projects by path instead of name.
Summary:
WorkspaceGenerator indexes groups and projects by their name. This
way it can add projects to existing groups instead of creating new
groups with the same name which would be very confusing. However
projects that have the same name are not merged in any way, instead
whichever project gets added later replaces the previous one. This
indexes projects by their path instead of their name. Note that
the name by which the projects are indexed is never used, as there
is no such attribute in the XML.

Test Plan:
CI
generate a workspace with two projects with the same name in the same grandparent directory.