Support using alternate Java compilers
Summary: This diff introduces a new "javac_jar" option in the Buck "tools" section.  When set, we use the compiler in that file instead of the normal system compiler. The support is not per-rule: the intent is that a single compiler is used for the project and "extra_options" is used to enable and disable nonstandard features of this compiler.

Test Plan: buck test