Fix FakeProjectFilesystem.getFilesUnderPath()
Summary:
`FakeProjectFilesystem.getFilesUnderPath()` is broken in a bunch
of ways currently. It had no tests, so it sadly regressed.

1) `ProjectFilesystem.walkRelativeFileTree()` is polymorphic, but the most
general version of the API which takes an `EnumSet<FileVisitOption>` was
private, so `FakeProjectFilesystem` could not override it.

2) `ProjectFilesystem.getFilesUnderPath()` incorrectly converted the
project-relative path to an absolute path before passing it to
`ProjectFilesystem.walkRelativeFileTree()`. That method requires a
project-relative path.

3) `FakeProjectFilesystem.getFilesUnderPath()` was implemented to
always return the empty set.

Test Plan:
New test added. Confirmed it failed until I fixed all above
issues.