Fix generated IntelliJ files for Windows Support
Summary:
Before this change, if you had a `prebuilt_jar` rule that wrapped a
`genrule`, we would include `:` in the name of the library because of
the reference to the target.  However, this would generate a file with
a colon in it, which is a no-no on Windows.

This change replaces `:` with `_` in our library name, much like we do
with `/`, `.`, and `-`.

Test Plan: `./bin/buck test //test/com/facebook/buck/java/intellij:java`