Add a setting to specify the name of the build files.
Summary:
Some build tools have similar build definition syntaxes to the one Buck uses.
With an option to choose the name of the build files and some shims defined
in the global include, it is possible to have a project use one set of build
files both for Buck and other build tools.
This adds that option.

Closes #238

Test Plan: CI, new test