Allow customizing productType in generated Xcode projects
Summary:
A user requested the ability to change the `productType`
field of generated Xcode projects.

Previously, we always guessed the `productType` based on the file
extension of the bundle or test. This doesn't work for Apple Watch
apps and extensions, which re-use the same `.app` and `.appex` file
extensions but have unique `productType`s in the Xcode project.

This diff adds a new optional field `xcode_product_type` to Apple
bundle and test rules which users can use to specify the `productType`
in the generated Xcode project.

Now that we have arbitrary product types, I changed
`com.facebook.buck.apple.xcodeproj.PBXTarget.ProductType` from an
`enum` to an abstract `Immutable` class with one `String` field.

Test Plan: Integration test included.