Work around the lack of hardlink support on Windows
Summary:
Windows doesn't have `os.link` so fall back to copying. This is what
the fallback code for os.link does anyways. The extra path check is because
`copyfile` will throw for this case but `os.link` throws an exception that the
existing code ignores.

Closes #247

Test Plan:
Python packager doesn't fall apart anymore with a missing attribute
for os.link. Files seems to make it into the pex.