Ignore the hierarchy of AppleSource
Summary:
When generating a project, ignore the tree hierarchy of `AppleSource`s. Instead
of copying their tree hierarchy into `GroupedSource`s, create the hierarchy of
the `GroupedSource`s based on the files location in the file system. This will
be followed by a diff that removes the option to nest `AppleSource`s and later
by a diff that completely removes `AppleSource`s and uses `SourcePath`s instead.

Test Plan:
  * Updated tests.
  * Generate a project, check that the groups in Xcode correspond to the files'
    physical locations.