allow ZipStep to include empty directory
Summary:
Right now ZipStep is not supporting zip empty dir, which introduce the problem for creating .aar since the `res/` dir is mandatory and it may be empty.
In this diff, when walking the file tree, store the directories and put into zipEntry. For a speical handling, if `zipEntry.getMethod() == ZipEntry.STORED`, skip zipping the directories.

Test Plan: arc unit