Generate GraphQL models during project generation.
Summary:
This changes Xcode project generation to pick up transitive GraphQL model
dependencies, merge them and build them before generating the project.
It does that by creating a new target graph with only the merged models
and their dependencies, and then executing a build with an action graph
created from that target graph. The merging logic creates for each
transitive dependent that needs a different combination of models a new
model with the same args as the component models, but with the union of
all the queries. This means the models have to have the same values for
args except for their queries to be merged successfully. The merged
models are shared between dependents as long as they're equivalent.

Test Plan:
CI
New unit tests.
Generate some projects with GraphQL dependencies, verify output in buck-out/gen/buck