buck test: don't build excluded tests
Summary:
Previously, when a user explicitly listed a test on the command which
got filtered out when using `--always_filter`, we'd still build that
target, even though we never ended up running.

This diff refactors the test code so that the `--all` and explicit
test modes use the same code path, where we explicitly construct the
target and action graphs and filter out tests correctly in both cases.

Test Plan: Added integration test to check cases.