Add ProjectFilesystem to AbstractSourcePath.
Summary:
`SourcePath` instances resolve to paths relative to the root of the repo. In
cross repo situations there isn't only one repo, so it's unknown which root
path source paths should resolve to. This adds a `ProjectFilesystem` field to
`SourcePath` so that it can be ultimately resolved to an absolute path instead
of a relative path, eliminating the ambiguity.

Test Plan: CI