Store per-file flags as a list of strings
Summary:
Instead of storing per-file flags as a single string in `AppleSources` and other
classes related to Apple targets, store them as a list of strings. This is more
compatible with other parts of our codebase. This change is not backward
compatible and will require updating existing apple targets.

Test Plan:
  * `buck test`.
  * Update an apple target to the new format and build it.