Normalize SourcePath handling in RuleKeys.
Summary:
In some code paths, the rule key for a `SourcePath` includes
the path to referenced files (eg when using immutable sorted
sets of sourcepaths) but in other cases only includes a hash
of the contents.

We are moving to a model where RuleKeys are calculated using
reflection, and it would be advantageous to handle all cases
in the same way.

Test Plan: buck test --all