Store headers separately from source in Apple constructor arg
Summary:
Introduce two new variables in `AppleNativeTargetDescriptionArg` (`headers` and
`exportedHeaders`) so headers are not mixed with sources, they don't have to use
per-file flags to distinguish exported headers from other headers and the
constructor arg is more similar to cxx targets.

This change is not backward compatible and existing apple targets will have to
be updated.

Test Plan:
  * `buck test`.
  * Update and compile an apple library.