Fix inconsistency with buck version UID and PEX deployment
Summary:
Previously, our PEX deployment of buck basically didn't have a buck
version UID, which meant a buck PEX and the in-repo buck that came
from the same buck source revision generated builds with different
rule keys.

This diff implements the correct buck version UID for buck.pex by
using a genrule help package it as a resource into the PEX.

Test Plan:
Ran a build with the in-repo buck and buck.pex.  Verified that rule
keys identical.