Drop support for android_binary()'s 'target' argument.
Summary:
The argument will still be accepted, but it will not be honored.

Previously, at the start of a Buck build, it would find all `android_binary()`
rules in the transitive deps of what was asked to be built and would
throw an exception if not all of the `target` arguments matched. This was a
poor and incorrect design decision.

This diff changes things so that the target, or Android SDK, is specified
at the repository level rather than the `android_binary()` level. This ensures
that all `android_binary()` rules in a repository can be built in the same
invocation of `buck build`.

This also defers the creation of the `AndroidPlatformTarget` until it is
needed whereas previously, an `AndroidPlatformTarget` was always created when
`buck build` was run, even if no Android targets were asked to be built.
This should clear up a lot of confusion for non-Android users of Buck who
would frequently see a message like the following displayed when doing a build,
even if it had nothing to do with Android:

    No Android platform target specified. Using default: Google Inc.:Google APIs:21

Now, the correct way to specify the repository's Android SDK is by doing so
in `.buckconfig`:

```
[android]
  target = Google Inc.:Google APIs:21
```

If no such target is specified, that of `AndroidPlatformTarget.DEFAULT_ANDROID_PLATFORM_TARGET`
will be used instead. (Currently, its value is "Google Inc.:Google APIs:21", but it will
change over time.) If the user does not have the specified target in his SDK, then an
exception will be thrown if it is required as part of the build.

There is still some outstanding work to do after this diff:

* Remove `AndroidPlatformTarget` from `ExecutionContext` and instead require that it is
passed from above into the `BuildRule`s that require it and ensure that the value of
`AndroidPlatformTarget.getName()` is included in the rule's `RuleKey`. The goal of passing
it from above is to make it clear when it needs to be incorporated into the `RuleKey`.
* Remove the use of `target` in Buck's own integration tests.
* Print a deprecation warning when `target` is specified, telling the user to
remove the `target` argument and add the appropriate information to `.buckconfig` instead.
* Once users have had sufficient time to update their Buck builds,
remove support for `target` in `android_binary()` altogether.

Test Plan: CI