Move a @VisibleForTesting constructor out of src/ and into test/.
Summary:
Previously, `CommandRunnerParamsForTesting` was a class under the `test/`
folder that was a subclass of `CommandRunnerParams`. Its constructor called
a superclass constructor that was `@VisibleForTesting`. The `@VisibleForTesting`
constructor existed only as a convenience to delegate to the real constructor.

This diff eliminates the unnecessary inheritance chain between
`CommandRunnerParamsForTesting` and `CommandRunnerParams`,
turning the former `@VisibleForTesting` constructor into a static creation
method in `CommandRunnerParamsForTesting`.

In a subsequent diff, the old `@VisibleForTesting` constructor will
need to take or create a `FakeBuckConfig`, which is not visible
from code under `src/`. Because this constructor was being used
exclusively from code under `test/`, it makes more sense for it
to live under `test/` as well where it is free to create a
`FakeBuckConfig` as needed.

Test Plan: Sandcastle builds.