Remove ExecutionContext#getAndroidPlatformTargetOptional().
Summary:
Now that the Android SDK is passed around via `Supplier<AndroidPlatformTarget>`,
there is no need to include the `Optional` qualifier. If the caller requests
an `AndroidPlatformTarget`, it should only be when he absolutely needs it.
Therefore, in the case when no such `AndroidPlatformTarget` is available,
a `NoAndroidSdkException` will be thrown by the `Supplier`, as appropriate.

The one place where the `Optional` qualifier was used somewhat legitimately
was in `Genrule`, but as commented in the code, that is a massive hack that
needs to go away. For now, we try/catch the call to `Supplier.get()` and
watch out for the `NoAndroidSdkException` explicitly.

Test Plan:
Introduced `BadAndroidConfigIntegrationTest`, which verifies that an incorrectly
configured Android SDK will not affect the building of a `java_library()` rule,
but will cause the building of an `android_library()` rule to fail.