Refactor ProjectCommand logic to create TargetGraph parser so it can be reused
Summary:
I'm working on refactoring the logic which builds the IDE-specific
target graph (including tests) from `ProjectCommand` so I can re-use it in
`TargetsCommand`.

The goal is to expose a `buck targets --show-target-hash ...` command which
can display a hash for an Xcode workspace rule including all dependencies
and tests. (I already implemented most of this except for the part to
include the tests).

This diff refactors the logic in `ProjectCommand` which creates the
`Parser` so I can re-use it from `TargetGraph`. In particular, I:

1) Changed from `BuildTarget` to `TargetNodeSpec`
2) Split out the `ProjectCommand` specific dependencies like
ignore paths and IDE type

The next step will be to move `ProjectGraphParser` creation out
into its own file which I can use from `TargetsCommand`.

Test Plan:
    % ant java-test -Dtest.class=ProjectCommand*Test
    java-test:
    [junit] Testsuite: com.facebook.buck.cli.ProjectCommandIntellijTest
    [junit] Tests run: 6, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 0.51 sec
    [junit]
    [junit] Testcase: testCreateTargetGraphWithoutTests took 0.376 sec
    [junit] Testcase: testCreateTargetGraphForSliceWithTests took 0.023 sec
    [junit] Testcase: testCreateTargetGraphForSliceWithoutTests took 0.025 sec
    [junit] Testcase: testCreateTargetGraphWithTests took 0.029 sec
    [junit] Testcase: testCreateTargetGraphForSmallSliceWithTests took 0.017 sec
    [junit] Testcase: testCreateTargetGraphForSmallSliceWithoutTests took 0.015 sec
    [junit] Testsuite: com.facebook.buck.cli.ProjectCommandXcodeTest
    [junit] Tests run: 6, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 0.159 sec
    [junit]
    [junit] Testcase: testCreateTargetGraphWithoutTests took 0.05 sec
    [junit] Testcase: testCreateTargetGraphForSliceWithTests took 0.029 sec
    [junit] Testcase: testCreateTargetGraphForSliceWithoutTests took 0.02 sec
    [junit] Testcase: testCreateTargetGraphWithTests took 0.022 sec
    [junit] Testcase: testCreateTargetGraphForSmallSliceWithTests took 0.018 sec
    [junit] Testcase: testCreateTargetGraphForSmallSliceWithoutTests took 0.015 sec

    BUILD SUCCESSFUL
    Total time: 2 seconds

Manually ran `buck project` in Xcode repository with and without arguments, confirmed it
kept previous behavior of only parsing needed `BUCK` files when arguments specified.