Remove unnecessary uses of TemporaryFolder.create().
Summary:
When a `TemporaryFolder` is created as a `@Rule`, its `create()`
method is called as part of its `before()` method, which it inherits
from `ExternalResource`. On the contrary, the Javadoc for `create()` is:

    for testing purposes only. Do not use.

As such, not calling it seems like the right thing to do.

I used Eclipse to find all references to `TemporaryFolder#create()` in
Buck and then deleted them.

Test Plan: Sandcastle