Fix `BuildInfoRecorder` to work on Windows
Summary:
We always tried to write out cache artifacts with the colon in the
name.  However, this is an illegal character on Windows, so we
never succeeded at writing the artifact out.  Hilariously, we
also logged success *and* failure of this.

This change also adds logging which would have made this issue
easier to spot in the first place.

Closes https://github.com/facebook/buck/issues/250

Test Plan:
* `ant test -D"test.class"="BuildInfoRecorderTest"`
* `ant test -D"test.class"="BuildInfoRecorderIntegrationTest"`