CachingBuildEngine: make cache "skipping" behavior configurable
Summary:
Currently, the `CachingBuildEngine` will "skip" attempting to fetch
builds from cache if an immediate dep was built locally.  This is an
optimization to avoid expensive artifacts fetches when it's unlikely
that we'll get a hit.

However, there some situations where this behavior is undesirable.
The first is with caches that are occasionally flaky.  If the cache
occasionally randomly drops an item or fails to connect this causes
a the entire build up the dep tree to be local, even if we could get
subsequent hits.

The second is when the case is really fast.  In this case, we get
little to no benefit avoiding it based on this heuristic.

In an attempt to explore alternatives to this, this diff makes the
behavior configurable by a setting in the `.buckconfig` which specifies
a "depth" for using this heuristic.  When set to `0`, this behavior
is disabled completely.  When set to some number greater than `0`,
we look a chain of consecutive local builds of at least this length
and skip the cache if we find it (the idea is that using a value such
as `2` provides a tradeoff between avoiding unnecessary cache fetches
and recovering from "random" misses).  The default is set to `1` which
keeps the current behavior.

Test Plan:
Ran in a repo using only a dircache.  Changed depth levels via .buckconfig
and ran builds.