Stopgap solution to eliminate a caching bug in python_binary() and python_test().
Summary:
Previously, in `PythonBinary`, `getInputsToCompareToOutput()` returned an empty list
and the cache information added by the `PythonPackageComponents` field did not contribute
any `Path`s to the `RuleKey.Builder`. This meant that no file contents were contributed
to the `RuleKey` (only file names), which meant that re-running a `python_test()` after
changing its code would have no effect because it the `RuleKey` would be unchanged and
the result (positive or negative) would simply be returned from the cache.

Test Plan:
Ran `buck build --keep-going <python-test-rule>` and verified that changing the source
of the test and running it again would show `BUILT_LOCALLY` instead of `MATCHING_RULE_KEY`.